// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PropertyPurpose = {
  venda: 'venda',
  aluguel: 'aluguel'
} as const

export type PropertyPurpose = (typeof PropertyPurpose)[keyof typeof PropertyPurpose]


export const PropertyType = {
  casa: 'casa',
  apartamento: 'apartamento',
  terreno: 'terreno',
  rural: 'rural',
  comercial: 'comercial'
} as const

export type PropertyType = (typeof PropertyType)[keyof typeof PropertyType]


export const PropertyCity = {
  corumba: 'corumba',
  ladario: 'ladario'
} as const

export type PropertyCity = (typeof PropertyCity)[keyof typeof PropertyCity]


export const PropertyStatus = {
  disponivel: 'disponivel',
  reservado: 'reservado',
  vendido: 'vendido',
  alugado: 'alugado'
} as const

export type PropertyStatus = (typeof PropertyStatus)[keyof typeof PropertyStatus]


export const UserRole = {
  admin: 'admin',
  user: 'user'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const LeadChannel = {
  whatsapp: 'whatsapp',
  telefone: 'telefone',
  email: 'email',
  formulario: 'formulario'
} as const

export type LeadChannel = (typeof LeadChannel)[keyof typeof LeadChannel]